# server/src/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shoreline Care OS API",
    "version": "6.2.0",
    "description": "Open-source dietary, clinical nutrition, and food operations platform API for care facilities. Generated from the Express route definitions by server/scripts/generate-openapi.mjs — do not edit by hand; run `npm run docs:generate` in server/."
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local Facility Gateway"
    },
    {
      "url": "https://api.shorelineops.com/api",
      "description": "Enterprise Cloud Sync"
    }
  ],
...
  "paths": {
    "/api/admin/audit": {
      "get": {
        "operationId": "get_admin_audit",
        "summary": "GET /api/admin/audit?limit=50&offset=0&userId=<uuid>",
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          }
        ],
        "description": "GET /api/admin/audit?limit=50&offset=0&userId=<uuid>\n\nAuthentication: JWT Bearer token; minimum role: admin (admits: admin).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "minRole": "admin"
        }
      }
    },
...
    "/api/audit": {
      "get": {
        "operationId": "get_audit",
        "summary": "GET /api/audit — admin only",
        "tags": [
          "audit"
        ],
        "description": "GET /api/audit — admin only\n\nAuthentication: JWT Bearer token; minimum role: admin (admits: admin).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "minRole": "admin"
        }
      },
      "post": {
        "operationId": "post_audit",
        "summary": "POST /api/audit — authenticated; actor taken from JWT only",
        "tags": [
          "audit"
        ],
        "description": "POST /api/audit — authenticated; actor taken from JWT only\n\nAuthentication: JWT Bearer token; any authenticated role.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditEventSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer"
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "post_auth_login",
        "summary": "POST /api/auth/login",
        "tags": [
          "auth"
        ],
        "description": "POST /api/auth/login\n\nPublic — no authentication required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/post_auth_login_Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [],
        "x-shoreline-auth": {
          "required": false,
          "mechanism": "none"
        }
      }
    },
...
    "/api/billing/webhook": {
      "post": {
        "operationId": "post_billing_webhook",
        "summary": "POST /api/billing/webhook",
        "tags": [
          "billing"
        ],
        "description": "POST /api/billing/webhook\nIngests Stripe payment lifecycle events.\nThe raw request body (preserved via express.raw in index.ts) is signature-checked\nwith STRIPE_WEBHOOK_SECRET before any processing. Forged events are rejected.\n\nAuthentication: Stripe webhook signature (stripe-signature header, STRIPE_WEBHOOK_SECRET).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request — validation failed"
          },
          "500": {
            "description": "Server error"
          },
          "503": {
            "description": "Service unavailable — integration not configured"
          }
        },
        "security": [
          {
            "stripeWebhookSignature": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "stripe-signature"
        }
      }
    },
...
    "/api/recipes": {
      "get": {
        "operationId": "get_recipes",
        "summary": "GET /api/recipes",
        "tags": [
          "recipes"
        ],
        "description": "GET /api/recipes\nList all master recipes with allergen tags and nutritional summary.\ncostProvenance is computed live from current vendor-catalog coverage so the\nSKU-matched vs estimated flag always reflects reality, never stale data.\n\nAuthentication: JWT Bearer token; any authenticated role.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer"
        }
      },
      "post": {
        "operationId": "post_recipes",
        "summary": "POST /api/recipes",
        "tags": [
          "recipes"
        ],
        "description": "POST /api/recipes\nCreate recipe, auto-detect allergens, and compute macro/micronutrients\n\nAuthentication: JWT Bearer token; minimum role: staff (admits: staff, server, activities, dietary, dietitian, frontdesk, manager, admin).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeBodySchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "minRole": "staff"
        }
      }
    },
...
    "/api/residents": {
      "get": {
        "operationId": "get_residents",
        "summary": "GET /api/residents[?q=<search>]",
        "tags": [
          "residents"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          }
        ],
        "description": "GET /api/residents[?q=<search>]\n\nAuthentication: JWT Bearer token; capabilities: residents.read or residents.serviceRead (admits: admin, manager, dietitian, frontdesk, dietary, activities, server, staff, readonly).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "admin",
            "manager",
            "dietitian",
            "frontdesk",
            "dietary",
            "activities",
            "server",
            "staff",
            "readonly"
          ],
          "capabilities": [
            "residents.read",
            "residents.serviceRead"
          ]
        }
      },
      "post": {
        "operationId": "post_residents",
        "summary": "POST /api/residents",
        "tags": [
          "residents"
        ],
        "description": "POST /api/residents\nAdmission always carries a clinical diet order: dietitian/manager only\n(residents.clinicalWrite, no admin wildcard). The capability gate IS the\nclinical authorization here, so replay after it is safe.\n\nAuthentication: JWT Bearer token; capabilities: residents.clinicalWrite (admits: dietitian, manager).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResidentSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "dietitian",
            "manager"
          ],
          "capabilities": [
            "residents.clinicalWrite"
          ]
        }
      }
    },
...
    "/api/timecard": {
      "get": {
        "operationId": "get_timecard",
        "summary": "GET /api/timecard — queryable punches",
        "tags": [
          "timecard"
        ],
        "parameters": [
          {
            "name": "badge_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          }
        ],
        "description": "GET /api/timecard — queryable punches\n\nAuthentication: JWT Bearer token; any authenticated role.",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer"
        },
        "x-shoreline-conditional": "Mounted only when ENABLE_TIMECARD_PLUGIN !== \"false\" (see src/index.ts)."
      }
    },
...
    "/api/trayruns": {
      "get": {
        "operationId": "get_trayruns",
        "summary": "GET /api/trayruns?serviceDate=&mealSlot=",
        "tags": [
          "trayruns"
        ],
        "parameters": [
          {
            "name": "mealSlot",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          },
          {
            "name": "serviceDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          }
        ],
        "description": "GET /api/trayruns?serviceDate=&mealSlot=\n\nAuthentication: JWT Bearer token; minimum role: staff (admits: staff, server, activities, dietary, dietitian, frontdesk, manager, admin).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request — validation failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "minRole": "staff"
        }
      },
      "post": {
        "operationId": "post_trayruns",
        "summary": "POST /api/trayruns — create a run explicitly (dispatch UI)",
        "tags": [
          "trayruns"
        ],
        "description": "POST /api/trayruns — create a run explicitly (dispatch UI)\n\nAuthentication: JWT Bearer token; minimum role: staff (admits: staff, server, activities, dietary, dietitian, frontdesk, manager, admin).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunBodySchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "minRole": "staff"
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by POST /api/auth/login. Role claim gates requireRole ranks."
      },
      "ehrWebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-EHR-Signature",
        "description": "HMAC-SHA256 of the raw body with EHR_WEBHOOK_SECRET."
      },
      "kioskApiSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Bearer <KIOSK_API_SECRET> for kiosk punch webhooks."
      },
      "stripeWebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "stripe-signature",
        "description": "Stripe webhook signature with STRIPE_WEBHOOK_SECRET."
      },
      "setupSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-setup-secret",
        "description": "SETUP_BOOTSTRAP_SECRET for first-time facility initialization."
      }
    },
    "schemas": {
      "AuditEventSchema": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "resourceId": {
            "type": "string"
          },
          "resourceType": {
            "type": "string"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "success",
              "failure"
            ]
          },
          "userAgent": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "action",
          "outcome"
        ]
      },
...
      "DateString": {
        "type": "string",
        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
      },
...
      "RecipeBodySchema": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "category": {
            "type": "string",
            "default": "Other"
          },
          "baseServings": {
            "type": "number",
            "minimum": 1,
            "default": 10
          },
          "prepTimeMins": {
            "type": "integer",
            "minimum": 0,
            "default": 15
          },
          "cookTimeMins": {
            "type": "integer",
            "minimum": 0,
            "default": 30
          },
          "haccpTempF": {
            "type": "number",
            "default": 165
          },
          "iddsiLevel": {
            "type": "integer",
            "minimum": 1,
            "maximum": 7,
            "default": 7
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeIngredientSchema"
            }
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeStepSchema"
            }
          },
          "notes": {
            "type": "string",
            "default": ""
          }
        },
        "required": [
          "name"
        ]
      },
      "RecipeIngredientSchema": {
        "type": "object",
        "properties": {
          "qty": {
            "type": "string",
            "minLength": 1
          },
          "item": {
            "type": "string",
            "minLength": 1
          },
          "vendorSku": {
            "type": "string"
          },
          "estimatedCost": {
            "type": "number"
          },
          "yieldPct": {
            "type": "number"
          }
        },
        "required": [
          "qty",
          "item"
        ]
      },
...
      "ResidentSchema": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "room": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Hospital",
              "LOA",
              "Passed Away"
            ],
            "default": "Active"
          },
          "dietType": {
            "type": "string",
            "default": "Regular"
          },
          "texture": {
            "type": "string",
            "default": "Regular"
          },
          "portionSize": {
            "type": "string",
            "default": "Regular"
          },
          "ensurePerDay": {
            "type": "integer",
            "minimum": 0,
            "default": 0
          },
          "allergies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "beverages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "birthdayMonth": {
            "type": "string"
          },
          "birthdayDay": {
            "type": [
              "integer",
              "null"
            ]
          },
          "servingLocation": {
            "type": "string",
            "default": "Dining Room"
          },
          "tableAssignment": {
            "type": "string",
            "default": ""
          },
          "likes": {
            "type": "string",
            "default": ""
          },
          "dislikes": {
            "type": "string",
            "default": ""
          },
          "specialInstructions": {
            "type": "string",
            "default": ""
          },
          "isNpo": {
            "type": "boolean"
          },
          "npoReason": {
            "type": "string"
          },
          "fluidRestrictionMl": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "dietEffectiveDate": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "room"
        ]
      },
...
      "RunBodySchema": {
        "type": "object",
        "properties": {
          "mealSlot": {
            "type": "string",
            "enum": [
              "MEAL_SLOTS"
            ]
          },
          "serviceDate": {
            "$ref": "#/components/schemas/DateString"
          },
          "wing": {
            "type": "string",
            "maxLength": 120,
            "default": ""
          },
          "notes": {
            "type": "string",
            "maxLength": 2000,
            "default": ""
          }
        },
        "required": [
          "mealSlot",
          "serviceDate"
        ]
      },
...
      "post_auth_login_Body": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  },
...
}